+++ src/tmb.cpp
@@ -25,6 +25,7 @@ Type objective_function<Type>::operator() ()
   DATA_VECTOR(population_t2);
   DATA_SPARSE_MATRIX(Lproj_hivpop);
   DATA_SPARSE_MATRIX(Lproj_incid);
+  DATA_SPARSE_MATRIX(Lproj_paed);
   DATA_SCALAR(projection_duration);
 
   // Design matrices
@@ -403,7 +404,7 @@ Type objective_function<Type>::operator() ()
   vector<Type> alpha_t2(invlogit(mu_alpha_t2));
 
   vector<Type> infections_t1t2((1 - exp(-lambda_t1 * projection_duration)) * (population_t1 - plhiv_t1));
-  vector<Type> plhiv_t2(Lproj_hivpop * plhiv_t1 + Lproj_incid * infections_t1t2);
+  vector<Type> plhiv_t2(Lproj_hivpop * plhiv_t1 + Lproj_incid * infections_t1t2 + Lproj_paed * plhiv_t1);
   
   vector<Type> rho_t2(plhiv_t2 / population_t2);
   vector<Type> prop_art_t2(rho_t2 * alpha_t2);