+++ NEWS.md
@@ -1,6 +1,7 @@
 # naomi 0.0.49
 
 * Properly age new infections between time 1 and time 2 based on ageing and incidecne trend from Spectrum.
+* Add new paediatric HIV infections betwen time 1 and time 2 proportional to HIV survivors by age and prevalence among women age 15-49 in Spectrum.
 
 # naomi 0.0.48